Move indexclient package contents into boltdb package
