+++ pkg/logcli/query/query_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	"github.com/grafana/loki/pkg/storage/config"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper"
-	"github.com/grafana/loki/pkg/storage/stores/shipper"
+	"github.com/grafana/loki/pkg/storage/stores/shipper/indexclient"
 	"github.com/grafana/loki/pkg/util/marshal"
 )
 
@@ -518,7 +518,7 @@ func TestLoadFromURL(t *testing.T) {
 	require.Error(t, err)
 	require.Nil(t, client)
 
-	conf.StorageConfig.BoltDBShipperConfig = shipper.Config{
+	conf.StorageConfig.BoltDBShipperConfig = indexclient.Config{
 		Config: indexshipper.Config{
 			SharedStoreType: config.StorageTypeFileSystem,
 		},
